services/render/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.0
pydantic-settings>=2.0.0
Pillow==10.4.0
PyJWT>=2.10.1
pytest==8.3.0
pytest-asyncio==0.24.0
